Moved `Utils` to base package and some methods to `JROMP` class
