+++ src/test/java/jromp/var/AtomicVariableTests.java
@@ -1,7 +1,6 @@
 package jromp.var;
 
 import jromp.JROMP;
-import jromp.utils.Utils;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -90,7 +89,7 @@ class AtomicVariableTests {
                      Variable<Integer> sum = variables.get("sum");
 
                      // The master thread will sleep for a while to end up with a different value
-                     if (Utils.isMaster(id)) {
+                     if (JROMP.isMaster(id)) {
                          try {
                              Thread.sleep(100);
                          } catch (InterruptedException e) {
